requirements.txt
fastapi==0.115.2 
uvicorn[standard]==0.30.6 
pydantic==2.9.2 
python-multipart==0.0.9 
deltalake==0.20.1 
pyarrow==17.0.0 
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0 
Faker==30.0.0
python-dotenv==1.1.1